requirements.txt
Django==4.1.3
python-decouple==3.6
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
pytest_django==4.5.2
factory_boy==3.2.1
flake8==6.0.0